+++ src/js/mstepper.js
@@ -14,15 +14,16 @@ class MStepper {
     */
    constructor(elem, options = {}) {
       this.stepper = elem;
-      this.options = {
-         firstActive: options.firstActive || 0,
-         linearStepsNavigation: options.linearStepsNavigation || true,
-         autoFocusInput: options.autoFocusInput || true,
-         showFeedbackPreloader: options.showFeedbackPreloader || true,
-         autoFormCreation: options.autoFormCreation || true,
-         validationFunction: options.validationFunction || null,
-         feedbackPreloader: options.feedbackPreloader || '<div class="preloader-wrapper active"> <div class="spinner-layer spinner-blue-only"> <div class="circle-clipper left"> <div class="circle"></div></div><div class="gap-patch"> <div class="circle"></div></div><div class="circle-clipper right"> <div class="circle"></div></div></div></div>'
-      };
+      this.options = Object.assign({
+         firstActive: 0,
+         linearStepsNavigation: true,
+         autoFocusInput: true,
+         showFeedbackPreloader: true,
+         autoFormCreation: true,
+         validationFunction: MStepper.defaultValidationFunction,
+         stepTitleNavigation: true,
+         feedbackPreloader: '<div class="preloader-wrapper active"> <div class="spinner-layer spinner-blue-only"> <div class="circle-clipper left"> <div class="circle"></div></div><div class="gap-patch"> <div class="circle"></div></div><div class="circle-clipper right"> <div class="circle"></div></div></div></div>'
+      }, options);
       this.classes = {
          HORIZONTALSTEPPER: 'horizontal',
          LINEAR: 'linear',
@@ -39,6 +40,8 @@ class MStepper {
       };
       this.events = {
          STEPCHANGE: new Event('stepchange'),
+         STEPOPEN: new Event('stepopen'),
+         STEPCLOSE: new Event('stepclose'),
          NEXTSTEP: new Event('nextstep'),
          PREVSTEP: new Event('prevstep'),
          STEPERROR: new Event('steperror'),
@@ -87,7 +90,8 @@ class MStepper {
          const submitButtons = step.querySelectorAll('button[type="submit"]');
          bindOrUnbind(nextBtns, 'click', _nextStepProxy, false);
          bindOrUnbind(prevBtns, 'click', _prevStepProxy, false);
-         bindOrUnbind(stepsTitle, 'click', _stepTitleClickHandler);
+         // Adding suggested feature in #62
+         if (options.stepTitleNavigation) bindOrUnbind(stepsTitle, 'click', _stepTitleClickHandler);
          // Prevents the tabbing issue (https://github.com/Kinark/Materialize-stepper/issues/49)
          if (inputs.length) bindOrUnbind(inputs[inputs.length - 1], 'keydown', tabbingDisabler);
          // Binds to the submit button an internal handler to manage validation
@@ -119,7 +123,7 @@ class MStepper {
     * @returns {HTMLElement} - The original received step.
     */
    _openAction = (step, cb, closeActiveStep = true, skipAutoFocus) => {
-      const { _slideDown, classes, getSteps, _closeAction, stepper, options } = this;
+      const { _slideDown, classes, getSteps, _closeAction, stepper, events, options } = this;
       // Gets the active step element
       const activeStep = getSteps().active.step;
       // If the active step is the same as the one that has been asked to be opened, returns the step
@@ -134,7 +138,7 @@ class MStepper {
          // Calls the slideDown private method if the stepper is vertical
          _slideDown(stepContent, classes.ACTIVESTEP, step, cb);
 
-         // Beginning of disabled autoFocusInput function due to issues with scroll
+         // Beginning of autoFocusInput
          if (!skipAutoFocus) {
             _slideDown(stepContent, classes.ACTIVESTEP, step, () => {
                // Gets the inputs from the nextStep to focus on the first one (temporarily disabled)
@@ -144,7 +148,7 @@ class MStepper {
                if (cb && typeof cb === 'function') cb();
             });
          }
-         // Enf of disabled autoFocusInput function due to issues with scroll
+         // Enf of autoFocusInput
 
       } else {
          // The stepper is running in horizontal mode
@@ -152,7 +156,14 @@ class MStepper {
          step.classList.add('active');
       }
       // If it was requested to close the active step as well, does it (default=true)
-      if (activeStep && closeActiveStep) _closeAction(activeStep);
+      if (activeStep && closeActiveStep) {
+         _closeAction(activeStep);
+         // We are changing steps, so dispatch the change event.
+         stepper.dispatchEvent(events.STEPCHANGE);
+      }
+      // Dispatch OPEN Event
+      stepper.dispatchEvent(events.STEPOPEN);
+
       return step;
    }
 
@@ -163,7 +174,7 @@ class MStepper {
     * @returns {HTMLElement} - The original received step.
     */
    _closeAction = (step, cb) => {
-      const { _slideUp, classes, stepper, _smartListenerUnbind, _smartListenerBind } = this;
+      const { _slideUp, classes, stepper, events, _smartListenerUnbind, _smartListenerBind } = this;
       // Gets the step content div inside the step
       const stepContent = step.getElementsByClassName(classes.STEPCONTENT)[0];
 
@@ -191,6 +202,8 @@ class MStepper {
          // Removes the class 'active' from the step, since all the animation is made by the CSS
          step.classList.remove('active');
       }
+      // Dispatch Event
+      stepper.dispatchEvent(events.STEPCLOSE);
       return step;
    }
 
@@ -262,8 +275,7 @@ class MStepper {
       // Opens the next one
       _openAction(nextStep, cb);
 
-      // Dispatches the events
-      stepper.dispatchEvent(events.STEPCHANGE);
+      // Dispatches the event
       stepper.dispatchEvent(events.NEXTSTEP);
    }
 
@@ -284,8 +296,7 @@ class MStepper {
       // Opens the previous step
       _openAction(prevStep, cb);
 
-      // Dispatches the events
-      stepper.dispatchEvent(events.STEPCHANGE);
+      // Dispatches the event
       stepper.dispatchEvent(events.PREVSTEP);
    }
 
@@ -296,16 +307,13 @@ class MStepper {
     * @returns {void}
     */
    openStep = (index, cb) => {
-      const { getSteps, _openAction, stepper, events, destroyFeedback } = this;
+      const { getSteps, _openAction, destroyFeedback } = this;
       const stepToOpen = getSteps().steps[index];
 
       // Destroyes the feedback preloader, if any
       destroyFeedback();
       // Opens the requested step
       _openAction(stepToOpen, cb);
-
-      // Dispatches the events
-      stepper.dispatchEvent(events.STEPCHANGE);
    }
 
    /**
@@ -405,18 +413,22 @@ class MStepper {
    activateStep = (elements, index) => {
       const { getSteps, _slideDown, stepper, _methodsBindingManager } = this;
       const { nodesIterator } = MStepper;
-      const currentSteps = getSteps();
-      const nextStep = currentSteps.steps[index];
+      const currentSteps = getSteps().steps;
+
+      // Checks if the steps will be added at the end or in the middle of the stepper
+      const before = currentSteps.length > index;
+      // Based on the previous check, sets the reference step
+      const referenceStep = before ? currentSteps[index] : currentSteps[currentSteps.length - 1];
 
       // Stores a let variable to return the right element after the activation
       let returnableElement = null;
       // Starts the checking of the elements parameter
       if (typeof elements === 'string') {
          // The element is in string format
-         // Insert it with the insertAdjacentHTML function
-         nextStep.insertAdjacentHTML('beforeBegin', elements);
+         // Insert it with the insertAdjacentHTML function (and trim the string to avoid errors)
+         referenceStep.insertAdjacentHTML(before ? 'beforeBegin' : 'afterEnd', elements.trim());
          // Defines the inserted element as the returnableElement
-         returnableElement = nextStep.previousSibling;
+         returnableElement = before ? referenceStep.previousSibling : referenceStep.nextSibling;
          // Activates (slideDown) the step
          _slideDown(returnableElement);
       } else if (Array.isArray(elements)) {
@@ -425,17 +437,21 @@ class MStepper {
          returnableElement = [];
          // Loops through the array
          elements.forEach(element => {
-            // Inserts each element with the insertAdjacentHTML function
-            nextStep.insertAdjacentHTML('beforeBegin', element);
+            // Inserts each element with the insertAdjacentHTML function (and trim the string to avoid errors)
+            referenceStep.insertAdjacentHTML(before ? 'beforeBegin' : 'afterEnd', element.trim());
+            // Gets the new added element
+            const addedStep = before ? referenceStep.previousSibling : referenceStep.nextSibling;
             // Adds each element to the returnableElement array
-            returnableElement.push(nextStep.previousSibling);
+            returnableElement.push(addedStep);
             // Activates (slideDown) each element
-            _slideDown(nextStep.previousSibling);
+            _slideDown(addedStep);
          });
       } else if (elements instanceof Element || elements instanceof HTMLCollection || elements instanceof NodeList) {
          // The element is an HTMLElement or an HTMLCollection
-         // Insert it/them with the insertBefore function and sets the returnableElement
-         returnableElement = stepper.insertBefore(elements, nextStep);
+         // Sets the rigth function to add the new steps
+         const rigthFunction = before ? stepper.insertBefore : stepper.appendChild;
+         // Insert it/them with the rigthFunction and sets the returnableElement
+         returnableElement = rigthFunction(elements, referenceStep);
          // If it's and HTMLElement, activates (slideDown) it, if it's an HTMLCollection, activates (slideDown) each of them
          if (elements instanceof Element) _slideDown(returnableElement); else nodesIterator(returnableElement, appendedElement => _slideDown(appendedElement));
       }
@@ -500,22 +516,26 @@ class MStepper {
 
       // Calls an animation frame to avoid async weird stuff
       requestAnimationFrame(() => {
-         // Prepare the element for animation
-         element.style.overflow = 'hidden';
-         element.style.paddingBottom = '0';
-         element.style.height = '0';
-         element.style.visibility = 'unset';
-         element.style.display = 'block';
-         // Calls another animation frame to wait for the previous changes to take effect
+         element.style.display = 'none';
          requestAnimationFrame(() => {
-            // Binds the "conclusion" function to the event 'transitionend'
-            this._smartListenerBind(element, 'transitionend', endSlideDown);
-            // Sets the final height to the element to trigger the transition
-            element.style.height = height;
-            // Removes the 'padding-bottom: 0' setted previously to trigger it too
-            element.style.removeProperty('padding-bottom');
-            // If a className for the slided element is required, add it
-            if (className) classElement.classList.add(className);
+            // Prepare the element for animation
+            element.style.overflow = 'hidden';
+            element.style.height = '0';
+            element.style.paddingBottom = '0';
+            element.style.visibility = 'unset';
+            element.style.display = 'block';
+            // Calls another animation frame to wait for the previous changes to take effect
+            requestAnimationFrame(() => {
+               // Binds the "conclusion" function to the event 'transitionend'
+               this._smartListenerBind(element, 'transitionend', endSlideDown);
+               // Sets the final height to the element to trigger the transition
+               element.style.height = height;
+               // Removes the 'padding-bottom: 0' setted previously to trigger it too
+               element.style.removeProperty('padding-bottom');
+               // element.style.paddingBottom = '0';
+               // If a className for the slided element is required, add it
+               if (className) classElement.classList.add(className);
+            });
          });
       });
       // Returns the original element to enable chain functions
@@ -723,30 +743,30 @@ class MStepper {
     * @returns {number} - The height without "px".
     */
    static getUnknownHeight(el) {
-      // Clones the element to insert it invisible
-      const clone = el.cloneNode(true);
-      // Defines some styles for it to be 100% invisible and unnoticeable
-      clone.style.position = 'fixed';
-      clone.style.display = 'block';
-      clone.style.top = '-999999px';
-      clone.style.left = '-999999px';
-      clone.style.height = 'auto';
-      clone.style.opacity = '0';
-      clone.style.zIndex = '-999999';
-      clone.style.pointerEvents = 'none';
-      // Rename the radio buttons in the cloned node as only 1 radio button is allowed to be selected with the same name in the DOM.
-      const radios = clone.querySelectorAll('[type="radio"]');
-      radios.forEach(radio => {
-         radio.name = "__" + radio.name + "__";
-      });
-      // Inserts it before the hidden element
-      const insertedElement = el.parentNode.insertBefore(clone, el);
+      // Spawns the hidden element in stealth mode
+      el.style.position = 'fixed';
+      el.style.display = 'block';
+      el.style.top = '-999999px';
+      el.style.left = '-999999px';
+      el.style.height = 'auto';
+      el.style.opacity = '0';
+      el.style.zIndex = '-999999';
+      el.style.pointerEvents = 'none';
       // Gets it's height
-      const height = insertedElement.offsetHeight;
-      // Removes it
-      el.parentNode.removeChild(insertedElement);
-      // Returns the height (without 'px')
+      const height = el.offsetHeight;
+      // Removes the stealth mode and hides the element again
+      MStepper.removeMultipleProperties(el, 'position display top left height opacity z-index pointer-events');
       return height;
+   }
+
+   /**
+    * Default validation function.
+    * @returns {boolean}
+    */
+   static defaultValidationFunction(stepperForm, activeStepContent) {
+      var inputs = activeStepContent.querySelectorAll('input, textarea, select');
+      for (let i = 0; i < inputs.length; i++) if (!inputs[i].checkValidity()) return false;
+      return true;
    }
 
    /**