+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 Date format: d/m/y
 
+### 3.0 (12/12/18)
+#### Features
+* Added `stepopen` and `stepclose` custom events;
+* Added `stepTitleNavigation` option as suggested in [#62](https://github.com/Kinark/Materialize-stepper/issues/62);
+* Added a default `validationFunction`.
+
+#### Fixes
+* Fixed [#54](https://github.com/Kinark/Materialize-stepper/issues/54) where you couldn't add a step to the end of the stepper;
+* Fixed [#58](https://github.com/Kinark/Materialize-stepper/issues/58) where specified `false` options were ignored;
+* Fixed [#60](https://github.com/Kinark/Materialize-stepper/issues/60) where `stepchange` event was not being fired when clicking on step-titles.
+* Refactored getUnknownHeight method to fix some bugs and optimize stuff;
+* Fixed animations issues;
+* Full migration to yarn.
+
+#### Docs
+* Added demos to API docs;
+* Updated sidebar with all sections.
+
 ### 3.0.0-beta.1.1.1 (28/11/18)
 * Added textarea when inputs need to be queried (fixes some issues);
 * Fix the listeners unbinding of steps that are being removed;